to_str method is working again for non-nested
