+++ catbird/action.py
@@ -1,7 +1,8 @@
 from .base import MooseBase
 
 class MooseAction(MooseBase):
-    _moose_action_params=[]
+
+    params_name="_moose_action_params"
 
     def __init__(self):
         print("I am an action")
